\name{satscanMapper.Version}
\alias{satscanMapper.Version}
\title{Function returns the release date and time of the satscanMapper package.  
 }
\description{
The satscanMapper.Version function is a support function that returns a string
set at the time the \var{satscanMapper} package was built and tested for release 
to CRAN.  The string contains the package name, version number and the date and
time the package was built.

"SaTScanMapper V1.0.2 - built 220823 at 8:08pm"

}
\usage{satscanMapper.Version()}
\value{A character vector contain the package name, version, and build data and time}
\author{ 
Jim Pearson and Linda Pickle of StatNet Consulting, LLC, Gaithersburg, MD 
}
